전체 폴더 구조 변경(components, pages 분리)
