+++ PandaMarket/src/pages/AddItem.jsx
@@ -1,0 +1,13 @@
+import { SubHeader } from "../layouts/Header";
+import AddItemWrapper from "../components/addItem/AddItemWrapper";
+
+const AddItem = () => {
+  return (
+    <>
+      <SubHeader />
+      <AddItemWrapper />
+    </>
+  );
+};
+
+export default AddItem;
