+++ PandaMarket/src/pages/Boards.jsx
@@ -1,0 +1,12 @@
+import { SubHeader } from "../layouts/Header";
+
+const Boards = () => {
+  return (
+    <>
+      <SubHeader />
+      <div>자유 게시판</div>
+    </>
+  );
+};
+
+export default Boards;
