--- Sprint-5-Mission/src/pages/items/Items.jsx
@@ -1,15 +1,0 @@
-import BestProduct from "./BestProduct";
-import AllProduct from "./AllProduct";
-import { SubHeader } from "../../Layouts/Header";
-
-const Items = () => {
-  return (
-    <>
-      <SubHeader />
-      <BestProduct />
-      <AllProduct />
-    </>
-  );
-};
-
-export default Items;
